+++ src/components/Accounts.tsx
@@ -27,8 +27,7 @@ import { useSourceTarget } from '../contexts/SourceTargetContextProvider';
 import { useAccounts } from '../hooks';
 import formatAccounts from '../util/formatAccounts';
 import getChainSS58 from '../util/getSS58';
-import Account from './Account';
-import SubHeader from './SubHeader';
+import { Account, SubHeader } from '.';
 
 interface Props {
   className?: string;
